restructured auth as well as base classes to not expose more than needed
