+++ Presentation.Web/Ninject/KernelBuilder.cs
@@ -409,6 +409,7 @@ namespace Presentation.Web.Ninject
             RegisterCommands<RemoveUserFromKitosCommandHandler>(kernel);
             RegisterCommands<RemoveOrganizationUnitRegistrationsCommandHandler>(kernel);
             RegisterCommands<AuthorizedUpdateOrganizationFromFKOrganisationCommandHandler>(kernel);
+            RegisterCommands<ValidateUserCredentialsCommandHandler>(kernel);
         }
 
         private void RegisterCommands<THandler>(IKernel kernel)
