+++ Presentation.Web/Presentation.Web.csproj
@@ -353,8 +353,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Controllers\API\V1\Auth\BaseAuthenticationController.cs" />
     <Compile Include="Controllers\API\V1\Auth\AuthorizeController.cs" />
+    <Compile Include="Controllers\API\V1\ExtendedApiController.cs" />
     <Compile Include="Models\API\V2\Request\SystemUsage\ArchivingCreationRequestDTO.cs" />
     <Compile Include="Models\API\V2\Request\SystemUsage\ArchivingUpdateRequestDTO.cs" />
     <Compile Include="Models\API\V2\Request\SystemUsage\IHasJournalPeriods.cs" />
